requirements.txt
altgraph==0.17.4
contourpy==1.3.2
cycler==0.12.1
fonttools==4.57.0
kiwisolver==1.4.8
matplotlib==3.10.1
numpy==2.2.5
objprint==0.3.0
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
pygame-ce==2.5.5
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pywin32-ctypes==0.2.3
setuptools==80.3.1
six==1.17.0
